[EDELIVERY-8336] put long instead of int for pMode id

+++ Domibus-MSH/src/main/java/eu/domibus/core/pmode/multitenancy/MultiDomainPModeProvider.java
@@ -299,7 +299,7 @@ public class MultiDomainPModeProvider extends PModeProvider {
     }
 
     @Override
-    public void removePMode(int id) {
+    public void removePMode(long id) {
         getCurrentPModeProvider().removePMode(id);
     }
